Update the decimal value for all. update the sending money gui view. update limit the text for card number update text bar name view

+++ app/src/main/java/com/example/springsophsoft/ui/AddCash/AddCashFragment.java
@@ -69,10 +69,10 @@ public class AddCashFragment extends Fragment {
         return root;
     }
 
-    public static int getCardBalance()
+    public static float getCardBalance()
     {
         String CardB = CardBalance;
-        int ConvertC = Integer.parseInt(CardB);
+        float ConvertC =  Float.parseFloat(CardB);
         return ConvertC;
     }
